requirements.txt
# requirements.txt

--index-url https://download.pytorch.org/whl/cu126

torch
torchvision
torchaudio
diffusers
transformers
accelerate
xformers
Pillow
qrcode
numpy